Add filter for active projects, fixes to filestructure and semantics + some styling to header and footer
