+++ src/Pages/page4.js
@@ -1,18 +1,12 @@
 import React from 'react';
+import GetMyProjects from '../services/notion/getFromNotion/projects/GetMyProjects'
 
 export default function Page4 () {
     return (
-        <div>
-        <main style={{
-            minHeight: '100vh'
-        }}>
-            <p style={{
-                marginLeft: '350px',
-                width: '60%',
-            }}>
-                Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-            </p>
-        </main>
+    <div style={{
+        minHeight: '100vh'
+    }}>
+        <GetMyProjects/>
     </div>
     )
 }
