+++ src/Pages/projects.js
@@ -1,15 +1,13 @@
 import React from 'react';
-import GetAllProjects from '../components/GetAllProjects';
+import GetAllProjects from '../services/notion/getFromNotion/projects/GetAllProjects';
 export default function Projects() {
     return (
-        <div>
-            <main style={{
+        <div style={{
             minHeight: '100vh'
         }}>
            <div>
             <GetAllProjects />
            </div>
-        </main>
         </div>
     );
 }
