--- src/components/Footer.js
@@ -1,20 +1,0 @@
-import React from 'react';
-
-const Footer = () => {
-  return (
-    <footer>
-      <p style={{
-                    textAlign: 'center',
-                    margin: '100px 0',
-                    backgroundColor: 'rgb(135, 135, 135)',
-                    paddingTop: '8px',
-                    paddingBottom: '8px',
-                    position: 'relative',
-                    bottom: '0',
-                    width: '100%',
-                }}>&copy; 2024 Elkineers. All rights reserved.</p>
-    </footer>
-  );
-};
-
-export default Footer;
