+++ src/components/Header/Header.js
@@ -1,0 +1,27 @@
+import React from 'react';
+import Login from '../../services/clerk/Login/Login'
+
+const Header = () => {
+  return (
+    <header
+        style={{
+            backgroundColor: 'teal',
+            color: 'white',
+            textAlign: 'center',
+            marginBottom: '100px',
+            display: 'flex',
+            justifyContent: 'space-between',
+            flexDirection: 'row'
+            }}>
+            <div></div>{/*Empty div to center h1*/}
+            <h1>ElkApp</h1>
+            <clerk style={{
+              margin: '15px'
+              }}>
+            <Login/>
+          </clerk>
+    </header>
+  );
+};
+
+export default Header;
